Add functionality to limit calls per consumer key (App) #207 - 2nd
